requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.6
redis==5.0.1
reportlab==4.0.9
openpyxl==3.1.2
pandas==2.2.0
matplotlib==3.8.3
seaborn==0.13.1
scikit-learn==1.3.2
gunicorn==21.2.0
whitenoise==6.6.0
django-crispy-forms>=2.3
crispy-bootstrap5==2024.10
requests==2.31.0
cryptography==42.0.3
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.6.0
django-mathfilters==1.0.0
